move funscriptplayer to own file
